+++ examples/example-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.4.146
+
+### Patch Changes
+
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.153
+
 ## 1.4.145
 
 ### Patch Changes